# docs/swagger.yaml
basePath: /api
definitions:
...
  payments.CheckoutResponse:
    properties:
      payment_id:
        type: string
      provider:
        type: string
      redirect_url:
        type: string
    type: object
  payments.MockCompleteRequest:
    properties:
      payment_id:
        type: string
    type: object
...
info:
  contact:
    email: aldoetobex@gmail.com
    name: Aldo Rifki Putra
  description: 'API for a mini legal marketplace: clients post cases, lawyers submit
    quotes, clients accept & pay, and lawyers access files via signed URLs.'
  title: Mini Legal Marketplace API
  version: "1.0"
paths:
...
  /checkout/{quoteID}:
    post:
      description: Create a Stripe Checkout Session using amount from DB
      parameters:
      - description: quote id (uuid)
        in: path
        name: quoteID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payments.CheckoutResponse'
      security:
      - BearerAuth: []
      summary: Create checkout (Stripe)
      tags:
      - payments
...
  /payments/mock/complete:
    post:
      consumes:
      - application/json
      description: 'Dev-only: finalize payment and mark a single quote as accepted'
      parameters:
      - description: Payment ID
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payments.MockCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            additionalProperties: true
            type: object
      summary: Complete payment (mock)
      tags:
      - payments
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"